# docs/接口文档/微友App-openapi.yaml
openapi: 3.0.3
info:
  title: 微友 App Backend API Draft
  version: 0.1.0-draft
  description: |
    基于当前后端控制器、需求文档与已生成接口文档整理的 OpenAPI 3.0 草稿。
    
    说明：
    - HTTP Base URL: `http://localhost:8080/api`
    - WebSocket URL: `ws://localhost:8081/ws/chat`
    - WebSocket 协议详情见 `docs/接口文档/微友App-后端接口文档.md`
    - 当前部分接口仍为 demo/stub/no-op 实现，字段结构可用于联调，业务结果未必代表最终生产行为
servers:
  - url: http://localhost:8080/api
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /wallet/bill/list:
    get:
      tags:
        - Wallet
      summary: 查询账单列表
      description: type、startDate、endDate 当前未真正参与查询过滤。
      operationId: walletBillList
      parameters:
        - in: query
          name: type
          required: false
          schema:
            type: string
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - in: query
          name: startDate
          required: false
          schema:
            type: string
        - in: query
          name: endDate
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageBillItemData'
  /wallet/transfer/create:
    post:
      tags:
        - Wallet
      summary: 创建转账
      description: 不允许给自己转账；金额必须大于 0；余额必须充足。
      operationId: walletTransferCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTransferResult'
  /wallet/red-packet/create:
    post:
      tags:
        - Wallet
      summary: 创建红包
      description: 红包总额必须大于等于红包个数。
      operationId: walletRedPacketCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedPacketCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRedPacketDetail'
  /wallet/red-packet/open:
    post:
      tags:
        - Wallet
      summary: 打开红包
      description: 重复领取直接返回原记录；过期、抢完或已结束时返回业务异常。
      operationId: walletRedPacketOpen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedPacketOpenRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRedPacketOpenResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponseBase:
      type: object
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        traceId:
          type: string
          example: 3f4b9d6e8a2c4f1b
...
    PageDataBase:
      type: object
      required:
        - pageNo
        - pageSize
        - total
        - hasMore
      properties:
        pageNo:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        hasMore:
          type: boolean
        nextCursor:
          type: string
          nullable: true
...
    BillItem:
      type: object
      properties:
        billId:
          type: integer
          format: int64
        transactionNo:
          type: string
        billType:
          type: string
        amountFen:
          type: integer
          format: int64
        incomeExpenseType:
          type: integer
          format: int32
          enum:
            - 1
            - 2
        billTitle:
          type: string
        billTime:
          type: string
    TransferCreateRequest:
      type: object
      required:
        - targetUserId
        - amountFen
      properties:
        targetUserId:
          type: integer
          format: int64
        amountFen:
          type: integer
          format: int64
        remark:
          type: string
    TransferResult:
      type: object
      properties:
        transactionId:
          type: integer
          format: int64
        transactionNo:
          type: string
        status:
          type: integer
          format: int32
        amountFen:
          type: integer
          format: int64
    RedPacketCreateRequest:
      type: object
      required:
        - amountFen
        - count
        - type
      properties:
        amountFen:
          type: integer
          format: int64
        count:
          type: integer
          format: int32
        type:
          type: integer
          format: int32
        greeting:
          type: string
        groupId:
          type: integer
          format: int64
          nullable: true
    RedPacketDetail:
      type: object
      properties:
        redPacketId:
          type: integer
          format: int64
        redPacketNo:
          type: string
        senderUserId:
          type: integer
          format: int64
        packetType:
          type: integer
          format: int32
        totalAmountFen:
          type: integer
          format: int64
        count:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
          enum:
            - 0
            - 1
            - 2
        expireAt:
          type: string
    RedPacketOpenRequest:
      type: object
      required:
        - redPacketId
      properties:
        redPacketId:
          type: integer
          format: int64
    RedPacketOpenResult:
      type: object
      properties:
        redPacketNo:
          type: string
        receiverUserId:
          type: integer
          format: int64
        receiveAmountFen:
          type: integer
          format: int64
        rankNo:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
...
    PageBillItemData:
      allOf:
        - $ref: '#/components/schemas/PageDataBase'
        - type: object
          properties:
            list:
              type: array
              items:
                $ref: '#/components/schemas/BillItem'
...
    ApiResponsePageBillItemData:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageBillItemData'
    ApiResponseTransferResult:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TransferResult'
    ApiResponseRedPacketDetail:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RedPacketDetail'
    ApiResponseRedPacketOpenResult:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RedPacketOpenResult'
...